Adds next for frontend rendering. Minor reorganization.
