+++ nodemon.json
@@ -1,5 +1,8 @@
 {
   "verbose": true,
-  "match": ["src/*"],
-  "ignore": ["dist/*"]
+  "watch": ["src/*"],
+  "ignore": ["dist/*"],
+  "env": {
+    "NODE_ENV": "development"
+  }
 }
