+++ package.json
@@ -2,26 +2,30 @@
   "name": "ozone",
   "version": "0.1.0",
   "private": true,
-  "main": "dist/app.js",
   "scripts": {
     "start": "node dist/app.js",
-    "dev": "nodemon -V -w src -e ts,tsx -x \"yarn build && yarn start\"",
+    "dev": "nodemon src/app.ts",
     "build": "tsc",
     "compose": "yarn install && yarn nodemon"
   },
   "dependencies": {
-    "@typegoose/typegoose": "^7.4.5",
     "cross-fetch": "^3.1.5",
     "discord.js": "12.5.3",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "node-cron": "^3.0.0"
+    "next": "^13.1.6",
+    "node-cron": "^3.0.0",
+    "prisma": "^4.9.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.7",
     "@types/node": "^14.6.0",
-    "nodemon": "^2.0.1-alpha.3",
+    "@types/react": "18.0.27",
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1",
     "tslint": "^6.1.3",
-    "typescript": "^4.0.2"
+    "typescript": "^4.9.5"
   }
 }
