+++ src/pages/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+/**
+ * The dashboard page
+ */
+function Dashboard() {
+  return (
+    <>Hello there!</>
+  )
+}
+
+export default Dashboard;
